% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{vectorize_propztest_for_property}
\alias{vectorize_propztest_for_property}
\title{Vectorize Proportion z-test for the Property data}
\usage{
vectorize_propztest_for_property(var, df, treatment_var = "treat")
}
\arguments{
\item{var}{the variable of which we want to compare the proportions}

\item{df}{dataframe that contains the data}

\item{treatment_var}{name of the treatment variable}
}
\value{
vectorized z-tests
}
\description{
Vectorize Proportion z-test for the Property data
}
