% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{vectorize_ttest_for_property}
\alias{vectorize_ttest_for_property}
\title{Vectorize T-test for the Property data}
\usage{
vectorize_ttest_for_property(x, property_df)
}
\arguments{
\item{x}{features of the Property dataframe}
}
\value{
vectorized t-tests
}
\description{
Vectorize T-test for the Property data
}
