% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{visualize_common_support}
\alias{visualize_common_support}
\title{Visualize common support from a propensity_model}
\usage{
visualize_common_support(propensity_model)
}
\arguments{
\item{propensity_model}{a propensity model}
}
\value{
a graph for common support
}
\description{
Visualize common support from a propensity_model
}
